requirements.txt
accelerate==1.12.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bce-python-sdk==0.9.60
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
fastapi==0.110.0
filelock==3.20.3
fsspec==2026.1.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
idna==3.11
imagesize==1.4.1
Jinja2==3.1.3
MarkupSafe==3.0.3
modelscope==1.34.0
mpmath==1.3.0
networkx==3.4.2
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.90
packaging==26.0
paddleocr==3.4.0
paddlex==3.4.1
pandas==2.2.3
pillow==10.3.0
prettytable==3.17.0
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.6.4
pydantic_core==2.16.3
pypdfium2==5.3.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.2
PyMuPDF
regex==2026.1.15
requests==2.32.5
ruamel.yaml==0.19.1
safetensors==0.7.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
starlette==0.36.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.29.0
wcwidth==0.5.3
render.yaml